Fix: 실제 사용자가 저장한 도서와 음악 개수, 작성한 서평과 음악 리뷰 개수가 반영되도록 수정

+++ roome/src/main/java/com/roome/domain/room/dto/RoomResponseDto.java
@@ -22,7 +22,7 @@ public class RoomResponseDto {
     private StorageLimitsDto storageLimits;
     private UserStorageDto userStorage;
 
-    public static RoomResponseDto from(Room room, int savedMusic, int savedBooks, int writtenReviews, int writtenMusicLogs) {
+    public static RoomResponseDto from(Room room, Long savedMusic, Long savedBooks, Long writtenReviews, Long writtenMusicLogs) {
         return RoomResponseDto.builder()
                 .roomId(room.getId())
                 .userId(room.getUser().getId())
